Problem 2 complete (cumpute_deriv)
Had to look up how to add to Tuples  using the notation: exTuple += (t,) where t is some variable. This was briefly mentioned in class but made much more sense in practice.

The reverse standard ordering of the elements in the tuple is very helpful. I would probabaly have to reverse it if the user entered the polynomial in standard form.